Updating README
readme should reflect the latest version 2.0.1 while defining the resource-type. With 2.0.0 I found a problem that in a pipeline having multiple jobs using the same or dedicated `metadata` resource-instances with `put` to for unique build url for specific jobs, was completely messed up and each job were getting `job-names` of other jobs. With tag 2.0.1, the problem away.